More easy moves that could have been done yesterday
