+++ rust/src/transformations/quantile_score_candidates/mod.rs
@@ -265,86 +265,3 @@ fn count_lt_eq<TI: PartialOrd>(x: &[TI], target: &TI) -> (usize, usize) {
 
 #[cfg(test)]
 mod test;
-
-#[cfg(all(test, feature = "derive"))]
-mod test_trans {
-    use crate::{
-        measurements::{make_report_noisy_max_gumbel, Optimize},
-        metrics::SymmetricDistance,
-    };
-
-    use super::*;
-
-    #[test]
-    fn test_scorer() -> Fallible<()> {
-        let candidates = vec![7, 12, 14, 72, 76];
-        let input_domain = VectorDomain::new(AtomDomain::default());
-        let input_metric = SymmetricDistance::default();
-        let trans =
-            make_quantile_score_candidates(input_domain, input_metric, candidates.clone(), 0.75)?;
-
-        let input_domain = VectorDomain::new(AtomDomain::default()).with_size(100);
-        let input_metric = SymmetricDistance::default();
-        let trans_sized =
-            make_quantile_score_candidates(input_domain, input_metric, candidates, 0.75)?;
-
-        let _scores = trans.invoke(&(0..100).collect())?;
-        let _scores_sized = trans_sized.invoke(&(0..100).collect())?;
-
-        // because alpha is .75, sensitivity is 1.5 (because not monotonic)
-        // granularity of quantile is .00001, so scores are integerized at a scale of 10000x
-        assert_eq!(trans.map(&1)?, 7500);
-
-        // alpha does not affect sensitivity- it's solely based on the size of the input domain
-        // using all of the range of the usize,
-        //     so scores can be scaled up by a factor of usize::MAX / 100 before being converted to integers and not overflow
-        // factor of 4 breaks into:
-        //   * a factor of 2 from non-monotonicity
-        //   * a factor of 2 from difference in score after moving one record from above to below a candidate
-        assert_eq!(trans_sized.map(&2)?, usize::MAX / 100 * 2);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_release() -> Fallible<()> {
-        let candidates = vec![7, 12, 14, 72, 76];
-        let input_domain = VectorDomain::new(AtomDomain::default());
-        let input_metric = SymmetricDistance::default();
-        let trans = make_quantile_score_candidates(input_domain, input_metric, candidates, 0.75)?;
-        let exp_mech = make_report_noisy_max_gumbel(
-            trans.output_domain.clone(),
-            trans.output_metric.clone(),
-            trans.map(&1)? as f64 * 2.,
-            Optimize::Min,
-        )?;
-
-        let quantile_meas = (trans >> exp_mech)?;
-        let idx = quantile_meas.invoke(&(0..100).collect())?;
-        println!("idx {:?}", idx);
-        assert!(quantile_meas.check(&1, &1.)?);
-        Ok(())
-    }
-
-    #[test]
-    fn test_release_sized() -> Fallible<()> {
-        let candidates = vec![7, 12, 14, 72, 76];
-        let input_domain = VectorDomain::new(AtomDomain::default()).with_size(100);
-        let input_metric = SymmetricDistance::default();
-        let trans_sized =
-            make_quantile_score_candidates(input_domain, input_metric, candidates, 0.75)?;
-        let exp_mech = make_report_noisy_max_gumbel(
-            trans_sized.output_domain.clone(),
-            trans_sized.output_metric.clone(),
-            trans_sized.map(&2)? as f64 * 2.,
-            Optimize::Min,
-        )?;
-
-        let quantile_sized_meas = (trans_sized >> exp_mech)?;
-        let idx = quantile_sized_meas.invoke(&(0..100).collect())?;
-        println!("idx sized {:?}", idx);
-        assert!(quantile_sized_meas.check(&1, &1.)?);
-
-        Ok(())
-    }
-}
